refactor: move AbstractGridResource to API
