+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/grid/AbstractPlatformGridResource.java
@@ -1,26 +1,27 @@
-package com.refinedmods.refinedstorage2.platform.common.internal.grid.view;
+package com.refinedmods.refinedstorage2.platform.api.grid;
 
 import com.refinedmods.refinedstorage2.api.grid.view.GridResourceAttributeKey;
 import com.refinedmods.refinedstorage2.api.grid.view.GridView;
 import com.refinedmods.refinedstorage2.api.resource.ResourceAmount;
 import com.refinedmods.refinedstorage2.api.storage.tracked.TrackedResource;
-import com.refinedmods.refinedstorage2.platform.api.grid.PlatformGridResource;
 
 import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 
-// TODO: Move to API
-public abstract class AbstractGridResource<T> implements PlatformGridResource {
+import org.apiguardian.api.API;
+
+@API(status = API.Status.STABLE, since = "2.0.0-milestone.3.0")
+public abstract class AbstractPlatformGridResource<T> implements PlatformGridResource {
     protected final ResourceAmount<T> resourceAmount;
     private final String name;
     private final Map<GridResourceAttributeKey, Set<String>> attributes;
     private boolean zeroed;
 
-    protected AbstractGridResource(final ResourceAmount<T> resourceAmount,
-                                   final String name,
-                                   final Map<GridResourceAttributeKey, Set<String>> attributes) {
+    protected AbstractPlatformGridResource(final ResourceAmount<T> resourceAmount,
+                                           final String name,
+                                           final Map<GridResourceAttributeKey, Set<String>> attributes) {
         this.resourceAmount = resourceAmount;
         this.name = name;
         this.attributes = attributes;
@@ -62,7 +63,7 @@ public abstract class AbstractGridResource<T> implements PlatformGridResource {
 
     @Override
     public String toString() {
-        return "AbstractGridResource{"
+        return "AbstractPlatformGridResource{"
             + "resourceAmount=" + resourceAmount
             + ", name='" + name + '\''
             + ", attributes=" + attributes
